backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.2.0
openai>=1.14.0
httpx>=0.27.0
python-dotenv==1.0.0
pydantic==2.6.0
python-multipart==0.0.6
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
# NEW DEPENDENCIES
python-docx==1.1.0
pypdf==4.0.1
openpyxl==3.1.2